Update 11.4
* Fix critical bug during update;
* Minor tweaks;

+++ code/r_central/osd/osd.cpp
@@ -55,7 +55,7 @@
 #include "../../base/ctrl_interfaces.h"
 #include "../../base/ctrl_settings.h"
 #include "../../base/hardware.h"
-#include "../../base/hw_procs.h"
+#include "../../base/hardware_procs.h"
 #include "../../base/utils.h"
 #include "../../base/tx_powers.h"
 
@@ -1109,14 +1109,14 @@ float osd_show_total_distance(float xPos, float yPos, float fScale)
    if ( pP->iUnits == prefUnitsImperial || pP->iUnits == prefUnitsFeets )
    {
       if ( _osd_convertKm(g_VehiclesRuntimeInfo[osd_get_current_data_source_vehicle_index()].headerFCTelemetry.total_distance/100.0/1000.0) > 1.0 )
-         snprintf(szBuff, sizeof(szBuff)/sizeof(szBuff[0]), "%s %.1f mi", szPrefix, (int)_osd_convertKm(g_VehiclesRuntimeInfo[osd_get_current_data_source_vehicle_index()].headerFCTelemetry.total_distance/100.0/1000.0));
+         snprintf(szBuff, sizeof(szBuff)/sizeof(szBuff[0]), "%s %.1f mi", szPrefix, (float)_osd_convertKm(g_VehiclesRuntimeInfo[osd_get_current_data_source_vehicle_index()].headerFCTelemetry.total_distance/100.0/1000.0));
       else
          snprintf(szBuff, sizeof(szBuff)/sizeof(szBuff[0]), "%s %d ft", szPrefix, (int)_osd_convertMeters(g_VehiclesRuntimeInfo[osd_get_current_data_source_vehicle_index()].headerFCTelemetry.total_distance/100));
    }
    else
    {
       if ( _osd_convertKm(g_VehiclesRuntimeInfo[osd_get_current_data_source_vehicle_index()].headerFCTelemetry.total_distance/100.0/1000.0) > 1.0 )
-         snprintf(szBuff, sizeof(szBuff)/sizeof(szBuff[0]), "%s %.1f km", szPrefix, (int)_osd_convertKm(g_VehiclesRuntimeInfo[osd_get_current_data_source_vehicle_index()].headerFCTelemetry.total_distance/100.0/1000.0));
+         snprintf(szBuff, sizeof(szBuff)/sizeof(szBuff[0]), "%s %.1f km", szPrefix, (float)_osd_convertKm(g_VehiclesRuntimeInfo[osd_get_current_data_source_vehicle_index()].headerFCTelemetry.total_distance/100.0/1000.0));
       else
          snprintf(szBuff, sizeof(szBuff)/sizeof(szBuff[0]), "%s %d m", szPrefix, (int)_osd_convertMeters(g_VehiclesRuntimeInfo[osd_get_current_data_source_vehicle_index()].headerFCTelemetry.total_distance/100));
    }
